+++ sktime_dl/classification/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 __all__ = [
+    "BaseDeepClassifier",
     "CNNClassifier",
-    "BaseClassifier",
 ]
-from sktime_dl.classification._classifier import BaseClassifier
+from sktime_dl.classification._classifier import BaseDeepClassifier
 from sktime_dl.classification._cnn import CNNClassifier
