+++ sktime_dl/classification/_cnn.py
@@ -4,7 +4,7 @@
 
 __author__ = "James Large"
 
-from sktime_dl.deeplearning.base.estimators import BaseDeepClassifier
+from sktime_dl.classification import BaseDeepClassifier
 from sktime_dl.networks._cnn import CNNNetwork
 from sktime_dl.utils import check_and_clean_data, \
     check_and_clean_validation_data
