+++ sktime_dl/deeplearning/base/estimators/__init__.py
@@ -6,6 +6,6 @@ __all__ = [
     "BaseDeepClassifier"
 ]
 
-from sktime_dl.classification._classifier import BaseDeepClassifier
+from sktime_dl.classification import BaseDeepClassifier
 from sktime_dl.networks._network import BaseDeepNetwork
 from sktime_dl.regression._regressor import BaseDeepRegressor
